fixed issue in Program.cs and moved main logic to App.axaml.cs to abide to Avalonia progrmmation standards
