+++ App.axaml
@@ -1,0 +1,7 @@
+<Application xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             x:Class="EasySave.App">
+    <Application.Styles>
+        <FluentTheme />
+    </Application.Styles>
+</Application>
